+++ pcs_test/tier1/legacy/test_rule.py
@@ -2,7 +2,8 @@
 from unittest import TestCase
 import xml.dom.minidom
 
-from pcs import rule
+from pcs import rule, utils
+from pcs.common.tools import Version
 from pcs_test.tools.assertions import ac, assert_xml_equal
 from pcs_test.tools.misc import (
     get_test_resource as rc,
@@ -15,7 +16,7 @@ from pcs_test.tools.pcs_runner import pcs
 # pylint: disable=invalid-name
 # pylint: disable=line-too-long
 
-empty_cib = rc("cib-empty.xml")
+empty_cib = rc("cib-empty-3.2.xml")
 
 
 class DateValueTest(TestCase):
@@ -328,6 +329,50 @@ class ParserTest(TestCase):
             str(self.parser.parse(["#uname", "eq", "integer", "-12345"])),
         )
         self.assertEqual(
+            "(eq (literal #uname) (number (literal 12345)))",
+            str(self.parser.parse(["#uname", "eq", "number", "12345"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 12.345)))",
+            str(self.parser.parse(["#uname", "eq", "number", "12.345"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 12345.)))",
+            str(self.parser.parse(["#uname", "eq", "number", "12345."])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal .12345)))",
+            str(self.parser.parse(["#uname", "eq", "number", ".12345"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 123e45)))",
+            str(self.parser.parse(["#uname", "eq", "number", "123e45"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 123E45)))",
+            str(self.parser.parse(["#uname", "eq", "number", "123E45"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 123e+45)))",
+            str(self.parser.parse(["#uname", "eq", "number", "123e+45"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 123E-45)))",
+            str(self.parser.parse(["#uname", "eq", "number", "123E-45"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal 12.34e5)))",
+            str(self.parser.parse(["#uname", "eq", "number", "12.34e5"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal +12.34e5)))",
+            str(self.parser.parse(["#uname", "eq", "number", "+12.34e5"])),
+        )
+        self.assertEqual(
+            "(eq (literal #uname) (number (literal -12.34e5)))",
+            str(self.parser.parse(["#uname", "eq", "number", "-12.34e5"])),
+        )
+        self.assertEqual(
             "(eq (literal #uname) (version (literal 1)))",
             str(self.parser.parse(["#uname", "eq", "version", "1"])),
         )
@@ -375,8 +420,16 @@ class ParserTest(TestCase):
             ["string", "#uname", "eq", "node1"],
         )
         self.assertSyntaxError(
-            "invalid integer value 'node1'",
-            ["#uname", "eq", "integer", "node1"],
+            "invalid integer value '123.45'",
+            ["#uname", "eq", "integer", "123.45"],
+        )
+        self.assertSyntaxError(
+            "invalid number value '123e45E67'",
+            ["#uname", "eq", "number", "123e45E67"],
+        )
+        self.assertSyntaxError(
+            "invalid number value '123.45.67'",
+            ["#uname", "eq", "number", "123.45.67"],
         )
         self.assertSyntaxError(
             "invalid version value 'node1'",
@@ -934,12 +987,116 @@ class ParserTest(TestCase):
             self.assertEqual(syntax_error, str(e))
 
 
+# already moved to pcs_test/tier0/lib/cib/rule/test_tools.py
+class HasNodeAttrExprWithTypeInteger(TestCase):
+    @staticmethod
+    def fixture_has_integer(rule_expression):
+        return rule.has_node_attr_expr_with_type_integer(
+            rule.RuleParser().parse(
+                rule.TokenPreprocessor().run(rule_expression)
+            )
+        )
+
+    def test_node_attr_no_type(self):
+        self.assertFalse(
+            self.fixture_has_integer(
+                [
+                    "(",
+                    "a",
+                    "eq",
+                    "A",
+                    "and",
+                    "b",
+                    "eq",
+                    "123",
+                    ")",
+                    "or",
+                    "a",
+                    "eq",
+                    "AA",
+                ]
+            )
+        )
+
+    def test_node_attr_no_integer(self):
+        self.assertFalse(
+            self.fixture_has_integer(
+                [
+                    "(",
+                    "a",
+                    "eq",
+                    "A",
+                    "and",
+                    "b",
+                    "eq",
+                    "number",
+                    "123",
+                    ")",
+                    "or",
+                    "a",
+                    "eq",
+                    "AA",
+                ]
+            )
+        )
+
+    def test_node_attr_integer(self):
+        self.assertTrue(
+            self.fixture_has_integer(
+                [
+                    "(",
+                    "a",
+                    "eq",
+                    "A",
+                    "and",
+                    "b",
+                    "eq",
+                    "integer",
+                    "123",
+                    ")",
+                    "or",
+                    "a",
+                    "eq",
+                    "AA",
+                ]
+            )
+        )
+
+    def test_node_attr_integer_not_date_expression(self):
+        self.assertTrue(
+            self.fixture_has_integer(["date", "gt", "integer", "123"])
+        )
+
+    def test_no_node_attr(self):
+        self.assertFalse(
+            self.fixture_has_integer(
+                [
+                    "(",
+                    "date-spec",
+                    "hours=1",
+                    "or",
+                    "defined",
+                    "pingd",
+                    ")",
+                    "or",
+                    "(",
+                    "date",
+                    "gt",
+                    "2014-06-26",
+                    "or",
+                    "date",
+                    "in_range",
+                    "2014-06-26",
+                    "to",
+                    "2014-07-26",
+                    ")",
+                ]
+            )
+        )
+
+
 # already moved to pcs_test/tier0/lib/cib/rule/test_parsed_to_cib.py
 class CibBuilderTest(TestCase):
-    def setUp(self):
-        self.parser = rule.RuleParser()
-        self.builder = rule.CibBuilder()
-
     # already moved to pcs_test/tier0/lib/cib/rule/test_parsed_to_cib.py
     def testSingleLiteralDatespec(self):
         self.assertExpressionXml(
@@ -1043,6 +1200,16 @@ class CibBuilderTest(TestCase):
             """,
         )
         self.assertExpressionXml(
+            ["#uname", "eq", "number", "12345"],
+            """
+<rsc_location id="location-dummy">
+    <rule id="location-dummy-rule">
+        <expression attribute="#uname" id="location-dummy-rule-expr" operation="eq" type="number" value="12345"/>
+    </rule>
+</rsc_location>
+            """,
+        )
+        self.assertExpressionXml(
             ["#uname", "eq", "integer", "12345"],
             """
 <rsc_location id="location-dummy">
@@ -1051,6 +1218,17 @@ class CibBuilderTest(TestCase):
     </rule>
 </rsc_location>
             """,
+        )
+        self.assertExpressionXml(
+            ["#uname", "eq", "integer", "12345"],
+            """
+<rsc_location id="location-dummy">
+    <rule id="location-dummy-rule">
+        <expression attribute="#uname" id="location-dummy-rule-expr" operation="eq" type="integer" value="12345"/>
+    </rule>
+</rsc_location>
+            """,
+            cib_file=rc("cib-empty-3.5.xml"),
         )
         self.assertExpressionXml(
             ["#uname", "eq", "version", "1.2.3"],
@@ -1154,7 +1332,7 @@ class CibBuilderTest(TestCase):
             """,
         )
         self.assertExpressionXml(
-            ["date", "gt", "integer", "12345"],
+            ["date", "gt", "number", "12345"],
             """
 <rsc_location id="location-dummy">
     <rule id="location-dummy-rule">
@@ -1476,17 +1654,20 @@ class CibBuilderTest(TestCase):
             """,
         )
 
-    def assertExpressionXml(self, rule_expression, rule_xml):
-        cib_dom = xml.dom.minidom.parse(empty_cib)
+    @staticmethod
+    def assertExpressionXml(rule_expression, rule_xml, cib_file=None):
+        cib_dom = xml.dom.minidom.parse(
+            cib_file if cib_file is not None else empty_cib
+        )
         constraints = cib_dom.getElementsByTagName("constraints")[0]
         constraint_el = constraints.appendChild(
             cib_dom.createElement("rsc_location")
         )
         constraint_el.setAttribute("id", "location-dummy")
         assert_xml_equal(
-            self.builder.build(
-                constraint_el, self.parser.parse(rule_expression)
-            ).parentNode.toprettyxml(indent="    "),
+            rule.CibBuilder(Version(*utils.getValidateWithVersion(cib_dom)))
+            .build(constraint_el, rule.RuleParser().parse(rule_expression))
+            .parentNode.toprettyxml(indent="    "),
             rule_xml.lstrip().rstrip(" "),
         )
 
@@ -1815,7 +1996,7 @@ class TokenPreprocessorTest(TestCase):
 
 
 class ExportAsExpressionTest(TestCase):
-    def test_success(self):
+    def test_success1(self):
         self.assertXmlExport(
             """
             <rule id="location-dummy-rule" score="INFINITY">
@@ -1826,6 +2007,8 @@ class ExportAsExpressionTest(TestCase):
             "#uname eq node1",
             "#uname eq string node1",
         )
+
+    def test_success2(self):
         self.assertXmlExport(
             """
             <rule id="location-dummy-rule" score="INFINITY">
@@ -1836,6 +2019,8 @@ class ExportAsExpressionTest(TestCase):
             "foo gt version 1.2.3",
             "foo gt version 1.2.3",
         )
+
+    def test_success3(self):
         self.assertXmlExport(
             """
 <rule boolean-op="or" id="complexRule" score="INFINITY">
@@ -1855,6 +2040,30 @@ class ExportAsExpressionTest(TestCase):
             """,
             '(date-spec hours=12-23 weekdays=1-5 and date in_range 2014-07-26 to duration months=1) or (foo gt version 1.2 and #uname eq "node3 4")',
             '(#uname eq string "node3 4" and foo gt version 1.2) or (date in_range 2014-07-26 to duration months=1 and date-spec hours=12-23 weekdays=1-5)',
+        )
+
+    def test_success_integer(self):
+        self.assertXmlExport(
+            """
+            <rule id="location-dummy-rule" score="INFINITY">
+                <expression attribute="foo" id="location-dummy-rule-expr"
+                    operation="gt" type="integer" value="123"/>
+            </rule>
+            """,
+            "foo gt integer 123",
+            "foo gt integer 123",
+        )
+
+    def test_success_number(self):
+        self.assertXmlExport(
+            """
+            <rule id="location-dummy-rule" score="INFINITY">
+                <expression attribute="foo" id="location-dummy-rule-expr"
+                    operation="gt" type="number" value="123"/>
+            </rule>
+            """,
+            "foo gt number 123",
+            "foo gt number 123",
         )
 
     @staticmethod
@@ -1877,19 +2086,7 @@ class ExportAsExpressionTest(TestCase):
         )
 
 
-class DomRuleAddTest(TestCase):
-    def setUp(self):
-        self.temp_cib = get_tmp_file("tier1_rule_dom_rule_add")
-        write_file_to_tmpfile(empty_cib, self.temp_cib)
-        output, returnVal = pcs(
-            self.temp_cib.name,
-            "resource create dummy1 ocf:heartbeat:Dummy".split(),
-        )
-        assert returnVal == 0 and output == ""
-
-    def tearDown(self):
-        self.temp_cib.close()
-
+class DomRuleAddXmlTest(TestCase):
     def test_success_xml(self):
         self.assertExpressionXml(
             ["#uname", "eq", "node1"],
@@ -1971,6 +2168,64 @@ class DomRuleAddTest(TestCase):
 </rsc_location>
             """,
         )
+        self.assertExpressionXml(
+            ["#uname", "eq", "integer", "12345"],
+            """
+<rsc_location id="location-dummy">
+    <rule id="location-dummy-rule" score="INFINITY">
+        <expression attribute="#uname" id="location-dummy-rule-expr" operation="eq" type="number" value="12345"/>
+    </rule>
+</rsc_location>
+            """,
+        )
+        self.assertExpressionXml(
+            ["#uname", "eq", "integer", "12345"],
+            """
+<rsc_location id="location-dummy">
+    <rule id="location-dummy-rule" score="INFINITY">
+        <expression attribute="#uname" id="location-dummy-rule-expr" operation="eq" type="integer" value="12345"/>
+    </rule>
+</rsc_location>
+            """,
+            cib_file=rc("cib-empty-3.5.xml"),
+        )
+
+    @staticmethod
+    def assertExpressionXml(rule_expression, rule_xml, cib_file=None):
+        cib_dom = xml.dom.minidom.parse(
+            cib_file if cib_file is not None else empty_cib
+        )
+        constraints = cib_dom.getElementsByTagName("constraints")[0]
+        constraint_el = constraints.appendChild(
+            cib_dom.createElement("rsc_location")
+        )
+        constraint_el.setAttribute("id", "location-dummy")
+        options, rule_argv = rule.parse_argv(rule_expression)
+        rule.dom_rule_add(
+            constraint_el,
+            options,
+            rule_argv,
+            Version(*utils.getValidateWithVersion(cib_dom)),
+        )
+        assert_xml_equal(
+            constraint_el.toprettyxml(indent="    "),
+            rule_xml.lstrip().rstrip(" "),
+        )
+
+
+class DomRuleAddTest(TestCase):
+    def setUp(self):
+        self.temp_cib = get_tmp_file("tier1_rule_dom_rule_add")
+        write_file_to_tmpfile(empty_cib, self.temp_cib)
+        output, returnVal = pcs(
+            self.temp_cib.name,
+            "resource create dummy1 ocf:heartbeat:Dummy".split(),
+        )
+        self.assertEqual(output, "")
+        self.assertEqual(returnVal, 0)
+
+    def tearDown(self):
+        self.temp_cib.close()
 
     @skip_unless_crm_rule()
     def test_success(self):
@@ -2173,18 +2428,3 @@ Location Constraints:
             "Error: id 'MyRule' is already in use, please specify another one\n",
         )
         self.assertEqual(1, returnVal)
-
-    @staticmethod
-    def assertExpressionXml(rule_expression, rule_xml):
-        cib_dom = xml.dom.minidom.parse(empty_cib)
-        constraints = cib_dom.getElementsByTagName("constraints")[0]
-        constraint_el = constraints.appendChild(
-            cib_dom.createElement("rsc_location")
-        )
-        constraint_el.setAttribute("id", "location-dummy")
-        options, rule_argv = rule.parse_argv(rule_expression)
-        rule.dom_rule_add(constraint_el, options, rule_argv)
-        assert_xml_equal(
-            constraint_el.toprettyxml(indent="    "),
-            rule_xml.lstrip().rstrip(" "),
-        )